backend/requirements-ims.txt
fastapi
uvicorn
prisma
pydantic[email]
python-jose[cryptography]
passlib
bcrypt==3.2.2
python-multipart
python-dotenv
aiofiles
